+++ pkg/costattribution/tracker_test.go
@@ -1,0 +1,199 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package costattribution
+
+import (
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus/testutil"
+	"github.com/prometheus/prometheus/model/labels"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/mimir/pkg/costattribution/testutils"
+	"github.com/grafana/mimir/pkg/mimirpb"
+)
+
+func TestTracker_hasSameLabels(t *testing.T) {
+	st := newTestManager().SampleTracker("user1")
+	assert.True(t, st.hasSameLabels([]string{"team"}), "Expected cost attribution labels mismatch")
+	ast := newTestManager().ActiveSeriesTracker("user1")
+	assert.True(t, ast.hasSameLabels([]string{"team"}), "Expected cost attribution labels mismatch")
+}
+
+func TestTracker_IncrementReceviedSamples(t *testing.T) {
+	tManager := newTestManager()
+	st := tManager.SampleTracker("user4")
+	t.Run("One Single Series in Request", func(t *testing.T) {
+		st.IncrementReceivedSamples(testutils.CreateRequest([]testutils.Series{{LabelValues: []string{"platform", "foo", "service", "dodo"}, SamplesCount: 3}}), time.Unix(10, 0))
+
+		expectedMetrics := `
+	# HELP cortex_received_attributed_samples_total The total number of samples that were received per attribution.
+	# TYPE cortex_received_attributed_samples_total counter
+	cortex_received_attributed_samples_total{platform="foo",tenant="user4",tracker="cost-attribution"} 3
+	`
+		assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(expectedMetrics), "cortex_received_attributed_samples_total"))
+	})
+	t.Run("Multiple Different Series in Request", func(t *testing.T) {
+		st.IncrementReceivedSamples(testutils.CreateRequest([]testutils.Series{
+			{LabelValues: []string{"platform", "foo", "service", "dodo"}, SamplesCount: 3},
+			{LabelValues: []string{"platform", "bar", "service", "yoyo"}, SamplesCount: 5},
+		}), time.Unix(20, 0))
+
+		expectedMetrics := `
+	# HELP cortex_received_attributed_samples_total The total number of samples that were received per attribution.
+	# TYPE cortex_received_attributed_samples_total counter
+	cortex_received_attributed_samples_total{platform="foo",tenant="user4",tracker="cost-attribution"} 6
+	cortex_received_attributed_samples_total{platform="bar",tenant="user4",tracker="cost-attribution"} 5
+	`
+		assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(expectedMetrics), "cortex_received_attributed_samples_total"))
+	})
+
+	t.Run("Multiple Series in Request with Same Labels", func(t *testing.T) {
+		st.IncrementReceivedSamples(testutils.CreateRequest([]testutils.Series{
+			{LabelValues: []string{"platform", "foo", "service", "dodo"}, SamplesCount: 3},
+			{LabelValues: []string{"platform", "foo", "service", "yoyo"}, SamplesCount: 5},
+		}), time.Unix(30, 0))
+
+		expectedMetrics := `
+	# HELP cortex_received_attributed_samples_total The total number of samples that were received per attribution.
+	# TYPE cortex_received_attributed_samples_total counter
+	cortex_received_attributed_samples_total{platform="foo",tenant="user4",tracker="cost-attribution"} 14
+	cortex_received_attributed_samples_total{platform="bar",tenant="user4",tracker="cost-attribution"} 5
+	`
+		assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(expectedMetrics), "cortex_received_attributed_samples_total"))
+	})
+}
+
+func TestTracker_CreateDelete(t *testing.T) {
+	tManager := newTestManager()
+	st := tManager.SampleTracker("user4")
+	ast := tManager.ActiveSeriesTracker("user4")
+
+	ast.Increment(labels.FromStrings("platform", "foo", "tenant", "user4", "team", "1"), time.Unix(1, 0))
+	ast.Increment(labels.FromStrings("platform", "foo", "tenant", "user4", "team", "2"), time.Unix(2, 0))
+	ast.Decrement(labels.FromStrings("platform", "foo", "tenant", "user4", "team", "3"))
+	st.IncrementReceivedSamples(testutils.CreateRequest([]testutils.Series{{LabelValues: []string{"platform", "foo", "team", "1"}, SamplesCount: 5}}), time.Unix(4, 0))
+	st.IncrementDiscardedSamples([]mimirpb.LabelAdapter{{Name: "platform", Value: "foo"}, {Name: "team", Value: "1"}}, 2, "sample-out-of-order", time.Unix(4, 0))
+	ast.Increment(labels.FromStrings("platform", "bar", "tenant", "user4", "team", "2"), time.Unix(6, 0))
+
+	expectedMetrics := `
+	# HELP cortex_discarded_attributed_samples_total The total number of samples that were discarded per attribution.
+    # TYPE cortex_discarded_attributed_samples_total counter
+    cortex_discarded_attributed_samples_total{platform="foo",reason="sample-out-of-order", tenant="user4",tracker="cost-attribution"} 2
+    # HELP cortex_ingester_attributed_active_series The total number of active series per user and attribution.
+    # TYPE cortex_ingester_attributed_active_series gauge
+	cortex_ingester_attributed_active_series{platform="bar",tenant="user4",tracker="cost-attribution"} 1
+    cortex_ingester_attributed_active_series{platform="foo",tenant="user4",tracker="cost-attribution"} 1
+    # HELP cortex_received_attributed_samples_total The total number of samples that were received per attribution.
+    # TYPE cortex_received_attributed_samples_total counter
+    cortex_received_attributed_samples_total{platform="foo",tenant="user4",tracker="cost-attribution"} 5
+	`
+
+	metricNames := []string{
+		"cortex_discarded_attributed_samples_total",
+		"cortex_received_attributed_samples_total",
+		"cortex_ingester_attributed_active_series",
+	}
+	assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(expectedMetrics), metricNames...))
+
+	// The purge only apply to the sample tracker.
+	assert.Equal(t, []string{"foo"}, st.inactiveObservations(time.Unix(5, 0)))
+	assert.NoError(t, tManager.purgeInactiveAttributionsUntil(time.Unix(5, 0)))
+
+	expectedMetrics = `
+	# HELP cortex_ingester_attributed_active_series The total number of active series per user and attribution.
+    # TYPE cortex_ingester_attributed_active_series gauge
+	cortex_ingester_attributed_active_series{platform="bar",tenant="user4",tracker="cost-attribution"} 1
+	cortex_ingester_attributed_active_series{platform="foo",tenant="user4",tracker="cost-attribution"} 1
+	`
+	assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(expectedMetrics), metricNames...))
+	tManager.deleteSampleTracker("user4")
+	tManager.deleteActiveTracker("user4")
+	assert.NoError(t, testutil.GatherAndCompare(tManager.reg, strings.NewReader(""), metricNames...))
+}
+
+func TestTracker_updateCounters(t *testing.T) {
+	st := newTestManager().SampleTracker("user3")
+	lbls1 := []mimirpb.LabelAdapter{{Name: "department", Value: "foo"}, {Name: "service", Value: "bar"}}
+	lbls2 := []mimirpb.LabelAdapter{{Name: "department", Value: "bar"}, {Name: "service", Value: "baz"}}
+	lbls3 := []mimirpb.LabelAdapter{{Name: "department", Value: "baz"}, {Name: "service", Value: "foo"}}
+
+	st.updateCountersWithLabelAdapter(lbls1, time.Unix(1, 0), 1, 0, nil)
+	assert.Equal(t, int64(0), st.overflowSince.Load(), "First observation, should not overflow")
+
+	st.updateCountersWithLabelAdapter(lbls2, time.Unix(2, 0), 1, 0, nil)
+	assert.Equal(t, int64(0), st.overflowSince.Load(), "Second observation, should not overflow")
+
+	st.updateCountersWithLabelAdapter(lbls3, time.Unix(3, 0), 1, 0, nil)
+	assert.Equal(t, int64(3), st.overflowSince.Load(), "Third observation, should overflow")
+
+	st.updateCountersWithLabelAdapter(lbls3, time.Unix(4, 0), 1, 0, nil)
+	assert.Equal(t, int64(3), st.overflowSince.Load(), "Fourth observation, should stay overflow")
+}
+
+func TestTracker_inactiveObservations(t *testing.T) {
+	// Setup the test environment: create a st for user1 with a "team" label and max cardinality of 5.
+	st := newTestManager().SampleTracker("user1")
+
+	// Create two observations with different last update timestamps.
+	observations := [][]mimirpb.LabelAdapter{
+		{{Name: "team", Value: "foo"}},
+		{{Name: "team", Value: "bar"}},
+		{{Name: "team", Value: "baz"}},
+	}
+
+	// Simulate samples discarded with different timestamps.
+	st.IncrementDiscardedSamples(observations[0], 1, "invalid-metrics-name", time.Unix(1, 0))
+	st.IncrementDiscardedSamples(observations[1], 2, "out-of-window-sample", time.Unix(12, 0))
+	st.IncrementDiscardedSamples(observations[2], 3, "invalid-metrics-name", time.Unix(20, 0))
+
+	// Ensure that two observations were successfully added to the tracker.
+	require.Len(t, st.observed, 3)
+
+	// Purge observations that haven't been updated in the last 10 seconds.
+	purged := st.inactiveObservations(time.Unix(0, 0))
+	require.Len(t, purged, 0)
+
+	purged = st.inactiveObservations(time.Unix(10, 0))
+	assert.ElementsMatch(t, []string{"foo"}, purged)
+
+	purged = st.inactiveObservations(time.Unix(15, 0))
+	assert.ElementsMatch(t, []string{"foo", "bar"}, purged)
+
+	// Check that the purged observation matches the expected details.
+	purged = st.inactiveObservations(time.Unix(25, 0))
+	assert.ElementsMatch(t, []string{"foo", "bar", "baz"}, purged)
+}
+
+func TestTracker_Concurrency(t *testing.T) {
+	m := newTestManager()
+	ast := m.ActiveSeriesTracker("user1")
+
+	var wg sync.WaitGroup
+	var i int64
+	for i = 0; i < 100; i++ {
+		wg.Add(1)
+		go func(i int64) {
+			defer wg.Done()
+			lbls := labels.FromStrings("team", string(rune('A'+(i%26))))
+			ast.Increment(lbls, time.Unix(i, 0))
+		}(i)
+	}
+	wg.Wait()
+
+	// Verify no data races or inconsistencies
+	assert.True(t, len(ast.observed) > 0, "Observed set should not be empty after concurrent updates")
+	assert.LessOrEqual(t, len(ast.observed), 2*ast.maxCardinality, "Observed count should not exceed 2 times of max cardinality")
+	assert.NotEqual(t, 0, ast.overflowSince.Load(), "Expected state to be Overflow")
+
+	expectedMetrics := `
+    # HELP cortex_ingester_attributed_active_series The total number of active series per user and attribution.
+    # TYPE cortex_ingester_attributed_active_series gauge
+	cortex_ingester_attributed_active_series{team="__overflow__",tenant="user1",tracker="cost-attribution"} 100
+`
+	assert.NoError(t, testutil.GatherAndCompare(m.reg, strings.NewReader(expectedMetrics), "cortex_ingester_attributed_active_series"))
+}